VideoAnonymizer.ObjectDetection/requirements.txt
fastapi
uvicorn[standard]
onnxruntime-gpu
opencv-python 
numpy 
pydantic
supervision
debugpy